# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateUserRequest": {
        "description": "Admin-created account. Password optional — omit it for an SSO-only staff\naccount, in which case `password` stays NULL and only Google can sign in.",
        "properties": {
          "account_type": {
            "default": "partner",
            "enum": [
              "staff",
              "partner"
            ],
            "title": "Account Type",
            "type": "string"
          },
          "company_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "designation": {
            "anyOf": [
              {
                "maxLength": 150,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Designation"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "employee_id": {
            "anyOf": [
              {
                "maxLength": 50,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Id"
          },
          "first_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "First Name",
            "type": "string"
          },
          "last_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Last Name",
            "type": "string"
          },
          "password": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "personal_email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personal Email"
          },
          "personal_mobile_number": {
            "anyOf": [
              {
                "maxLength": 30,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personal Mobile Number"
          },
          "role_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Role Ids",
            "type": "array"
          },
          "status": {
            "default": "INACTIVE",
            "enum": [
              "INACTIVE",
              "ACTIVE",
              "SUSPENDED"
            ],
            "title": "Status",
            "type": "string"
          },
          "timezone_preference": {
            "default": "Asia/Kolkata",
            "title": "Timezone Preference",
            "type": "string"
          }
        },
        "required": [
          "first_name",
          "last_name",
          "email"
        ],
        "title": "CreateUserRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PaginatedActivity": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActivityEntry"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "per_page": {
            "title": "Per Page",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "per_page",
          "pages"
        ],
        "title": "PaginatedActivity",
        "type": "object"
      },
      "PaginatedUsers": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "per_page": {
            "title": "Per Page",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "per_page",
          "pages"
        ],
        "title": "PaginatedUsers",
        "type": "object"
      },
...
      "UserDetailResponse": {
        "properties": {
          "account_type": {
            "title": "Account Type",
            "type": "string"
          },
          "auth_provider": {
            "title": "Auth Provider",
            "type": "string"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "can_approve": {
            "default": false,
            "title": "Can Approve",
            "type": "boolean"
          },
          "can_delete": {
            "default": false,
            "title": "Can Delete",
            "type": "boolean"
          },
          "can_edit": {
            "default": false,
            "title": "Can Edit",
            "type": "boolean"
          },
          "can_toggle_status": {
            "default": false,
            "title": "Can Toggle Status",
            "type": "boolean"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "designation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Designation"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "email_verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Verified At"
          },
          "employee_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Employee Id"
          },
          "failed_login_attempts": {
            "title": "Failed Login Attempts",
            "type": "integer"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "initials": {
            "title": "Initials",
            "type": "string"
          },
          "last_login_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "last_login_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login Ip"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "locked_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locked Until"
          },
          "personal_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personal Email"
          },
          "personal_mobile_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Personal Mobile Number"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/RoleSummary"
            },
            "title": "Roles",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "timezone_preference": {
            "title": "Timezone Preference",
            "type": "string"
          },
          "two_factor_enabled": {
            "default": false,
            "title": "Two Factor Enabled",
            "type": "boolean"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "full_name",
          "initials",
          "avatar_url",
          "designation",
          "company_name",
          "account_type",
          "status",
          "auth_provider",
          "last_login_at",
          "created_at",
          "roles",
          "first_name",
          "last_name",
          "employee_id",
          "personal_mobile_number",
          "personal_email",
          "timezone_preference",
          "email_verified_at",
          "failed_login_attempts",
          "locked_until",
          "last_login_ip",
          "updated_at"
        ],
        "title": "UserDetailResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Partner Marketplace API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/activity": {
      "get": {
        "description": "The audit trail, newest first.\n\nSorted by `id`, not `created_at`: rows written inside one transaction can share\na timestamp, and an unstable sort would let a row appear on two consecutive\npages or on neither.",
        "operationId": "list_activity_api_v1_activity_get",
        "parameters": [
          {
            "description": "'auth' | 'default'",
            "in": "query",
            "name": "log_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "'auth' | 'default'",
              "title": "Log Name"
            }
          },
          {
            "in": "query",
            "name": "event",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event"
            }
          },
          {
            "description": "'User' | 'Role'",
            "in": "query",
            "name": "subject_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "'User' | 'Role'",
              "title": "Subject Type"
            }
          },
          {
            "in": "query",
            "name": "subject_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject Id"
            }
          },
          {
            "description": "Filter to one actor",
            "in": "query",
            "name": "causer_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one actor",
              "title": "Causer Id"
            }
          },
          {
            "description": "Substring of the description",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Substring of the description",
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "description": "Drop rows with no human causer (automation)",
            "in": "query",
            "name": "hide_system",
            "required": false,
            "schema": {
              "default": false,
              "description": "Drop rows with no human causer (automation)",
              "title": "Hide System",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Per Page",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedActivity"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Activity",
        "tags": [
          "activity"
        ]
      }
    },
    "/api/v1/activity/events": {
      "get": {
        "description": "Event names actually present in the trail, for the filter dropdown.\n\nRead from the data rather than a hardcoded list, so an event added by a future\ncall site appears without anyone remembering to register it — and one that has\nnever occurred does not clutter the filter.",
        "operationId": "list_events_api_v1_activity_events_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response List Events Api V1 Activity Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Events",
        "tags": [
          "activity"
        ]
      }
    },
    "/api/v1/activity/export": {
      "get": {
        "description": "Stream the trail as CSV — the first thing anyone asks for in a real review.\n\n**Streamed, not built in memory.** This is the one read with no upper bound:\n\"everything for the audit\" is the whole point, and materialising a year of rows\nto assemble a response is the request that runs the process out of memory.\n\nOldest first, unlike the paginated view: a file is read top to bottom as a\nchronology, where a screen is read newest-first.\n\n`properties` is emitted as compact JSON in one column. Flattening it into\ncolumns is impossible — the shape differs per event — and truncating it would\nsilently drop the before/after diff that makes an export worth having.",
        "operationId": "export_activity_api_v1_activity_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "log_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Log Name"
            }
          },
          {
            "in": "query",
            "name": "event",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Activity",
        "tags": [
          "activity"
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "operationId": "list_users_api_v1_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "account_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Type"
            }
          },
          {
            "in": "query",
            "name": "role_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role Id"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "created_at",
              "title": "Sort By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort_order",
            "required": false,
            "schema": {
              "default": "desc",
              "pattern": "^(asc|desc)$",
              "title": "Sort Order",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "default": 15,
              "maximum": 100,
              "minimum": 1,
              "title": "Per Page",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUsers"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "create_user_api_v1_users_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "access_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create User",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}